moves reprocess button menu to expired job tag
