+++ src/app/components/results-menu/scene-files/scene-file/scene-file.component.html
@@ -108,10 +108,7 @@
     <mat-menu #expiredJobs ='matMenu'>
 
     <button (click)="queueExpiredHyp3Job()" mat-menu-item>
-      Add Expired {{ product.metadata.job.job_type }} job to On Demand Queue
-    </button>
-    <button (click)="onReviewExpiredJob()" [disabled]="remaining < 1" [matTooltip]="remaining > 0 ? '' : 'No more jobs available for this month'" mat-menu-item>
-      Re-Process Expired {{ product.metadata.job.job_type }} Job
+      Add job
     </button>
   </mat-menu>
   </div>
