+++ src/app/components/shared/hyp3-job-status-badge/hyp3-job-status-badge.component.html
@@ -1,6 +1,12 @@
 <mat-chip-list aria-label="Job Status">
-  <mat-chip *ngIf="isExpired(job)" color="primary">Expired</mat-chip>
+  <mat-chip [ngClass]="clickable" [matMenuTriggerFor]='expiredJobMenu' *ngIf="isExpired(job)" color="primary">Expired...</mat-chip>
   <mat-chip *ngIf="isPending(job)" color="secondary">Pending</mat-chip>
   <mat-chip *ngIf="isFailed(job)" color="accent">Failed</mat-chip>
   <mat-chip *ngIf="isRunning(job)" color="secondary">Running</mat-chip>
 </mat-chip-list>
+
+<mat-menu #expiredJobMenu = "matMenu">
+<button (click)="onReviewExpiredJob()" [disabled]="remaining < 1" mat-menu-item>
+    Resubmit
+</button>
+</mat-menu>
