+++ src/app/components/shared/hyp3-job-status-badge/hyp3-job-status-badge.component.scss
@@ -1,0 +1,3 @@
+.clickable {
+  cursor: pointer;
+}
