+++ src/app/components/shared/hyp3-job-status-badge/hyp3-job-status-badge.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatChipsModule } from '@angular/material/chips';
 
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatButtonModule } from '@angular/material/button';
 import { Hyp3JobStatusBadgeComponent } from './hyp3-job-status-badge.component';
 
 
@@ -13,6 +16,9 @@ import { Hyp3JobStatusBadgeComponent } from './hyp3-job-status-badge.component';
   imports: [
     CommonModule,
     MatChipsModule,
+    MatFormFieldModule,
+    MatMenuModule,
+    MatButtonModule
   ],
   exports: [
     Hyp3JobStatusBadgeComponent
