+++ src/app/components/shared/on-demand-add-menu/on-demand-add-menu.component.html
@@ -45,23 +45,9 @@
         (click)="queueAllOnDemand(prodType.products, byJobType.jobType)"
         [disabled]="" mat-menu-item>
 
-        Add {{isExpired ? 'Expired' : ''}}  {{ isExpired ? '' : prodType.products.length }} {{ prodType.productType }}
+        Add  {{ isExpired ? '' : prodType.products.length}} {{ isExpired ? '' : prodType.productType }}
         {{ byJobType.jobType.numProducts === 2 ? 'pair' : 'job' }}{{ prodType.products.length > 1 ? 's' : '' }}
       </button>
-
-      <div *ngIf="isExpired && expiredJobs.job_type === prodType.productType">
-        <button
-          [disabled]="!expiredJobs && remaining < 1"
-          (click)="onReviewExpiredJob()"
-          mat-menu-item>
-          Re-Process Expired {{expiredJobs.job_type}} {{ byJobType.jobType.numProducts === 2 ? 'pair' : 'job' }}
-
-          <mat-icon
-            class="clickable" style="margin-left: 10px;">
-              icon
-          </mat-icon>
-        </button>
-      </div>
     </div>
 
     <button (click)="onOpenCreateSubscription()" mat-menu-item>Create Subscription</button>
